entity/diff: Entity 和 TqDiff 分别在不同的子模块，消除循环依赖
