+++ tqsdk/entity.py
@@ -1,0 +1,38 @@
+#!usr/bin/env python3
+#-*- coding:utf-8 -*-
+__author__ = 'yanqiong'
+
+import copy
+import weakref
+from collections.abc import MutableMapping
+
+
+class Entity(MutableMapping):
+    def _instance_entity(self, path):
+        self._path = path
+        self._listener = weakref.WeakSet()
+
+    def __setitem__(self, key, value):
+        return self.__dict__.__setitem__(key, value)
+
+    def __delitem__(self, key):
+        return self.__dict__.__delitem__(key)
+
+    def __getitem__(self, key):
+        return self.__dict__.__getitem__(key)
+
+    def __iter__(self):
+        return iter({k: v for k, v in self.__dict__.items() if not k.startswith("_")})
+
+    def __len__(self):
+        return len({k: v for k, v in self.__dict__.items() if not k.startswith("_")})
+
+    def __str__(self):
+        return str({k: v for k, v in self.__dict__.items() if not k.startswith("_")})
+
+    def __repr__(self):
+        return '{}, D({})'.format(super(Entity, self).__repr__(),
+                                  {k: v for k, v in self.__dict__.items() if not k.startswith("_")})
+
+    def copy(self):
+        return copy.copy(self)
