+++ tqsdk/objs.py
@@ -3,39 +3,9 @@
 __author__ = 'chengzhi'
 
 import copy
-import weakref
-from collections.abc import MutableMapping
 
-
-class Entity(MutableMapping):
-    def _instance_entity(self, path):
-        self._path = path
-        self._listener = weakref.WeakSet()
-
-    def __setitem__(self, key, value):
-        return self.__dict__.__setitem__(key, value)
-
-    def __delitem__(self, key):
-        return self.__dict__.__delitem__(key)
-
-    def __getitem__(self, key):
-        return self.__dict__.__getitem__(key)
-
-    def __iter__(self):
-        return iter({k: v for k, v in self.__dict__.items() if not k.startswith("_")})
-
-    def __len__(self):
-        return len({k: v for k, v in self.__dict__.items() if not k.startswith("_")})
-
-    def __str__(self):
-        return str({k: v for k, v in self.__dict__.items() if not k.startswith("_")})
-
-    def __repr__(self):
-        return '{}, D({})'.format(super(Entity, self).__repr__(),
-                                  {k: v for k, v in self.__dict__.items() if not k.startswith("_")})
-
-    def copy(self):
-        return copy.copy(self)
+from tqsdk.entity import Entity
+from tqsdk.diff import TqDiff
 
 
 class Quote(Entity):
@@ -427,7 +397,7 @@ class Position(Entity):
 
         :return: dict, 其中每个元素的key为委托单ID, value为 :py:class:`~tqsdk.objs.Order`
         """
-        tdict = self._api._get_obj(self._api._data, ["trade", self._api._account._account_id, "orders"])
+        tdict = TqDiff._get_obj(self._api._data, ["trade", self._api._account._account_id, "orders"])
         fts = {order_id: order for order_id, order in tdict.items() if (not order_id.startswith(
             "_")) and order.instrument_id == self.instrument_id and order.exchange_id == self.exchange_id and order.status == "ALIVE"}
         return fts
@@ -511,7 +481,7 @@ class Order(Entity):
 
         :return: 当委托单部分成交或全部成交时, 返回成交部分的平均成交价. 无任何成交时, 返回 nan
         """
-        tdict = self._api._get_obj(self._api._data, ["trade", self._api._account._account_id, "trades"])
+        tdict = TqDiff._get_obj(self._api._data, ["trade", self._api._account._account_id, "trades"])
         sum_volume = sum([trade.volume for trade_id, trade in tdict.items() if
                           (not trade_id.startswith("_")) and trade.order_id == self.order_id])
         if sum_volume == 0:
@@ -527,7 +497,7 @@ class Order(Entity):
 
         :return: dict, 其中每个元素的key为成交ID, value为 :py:class:`~tqsdk.objs.Trade`
         """
-        tdict = self._api._get_obj(self._api._data, ["trade", self._api._account._account_id, "trades"])
+        tdict = TqDiff._get_obj(self._api._data, ["trade", self._api._account._account_id, "trades"])
         fts = {trade_id: trade for trade_id, trade in tdict.items() if
                (not trade_id.startswith("_")) and trade.order_id == self.order_id}
         return fts
